consolidate TallyPopulation
