+++ src/NatCruise.Core.Test/Data/CuttingUnitDatastore_Test.cs
@@ -1,0 +1,20 @@
+﻿using FluentAssertions;
+using NatCruise.Data;
+using NatCruise.Cruise.Models;
+using NatCruise.Models;
+using NatCruise.Test;
+using System.Linq;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace NatCruise.Test.Data
+{
+    public class CuttingUnitDatastore_Test : TestBase
+    {
+        public CuttingUnitDatastore_Test(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        
+    }
+}
