+++ src/NatCruise.Core.Test/NatCruise.Core.Test.csproj
@@ -21,6 +21,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\NatCruise.Core\NatCruise.Core.csproj" />
+		<ProjectReference Include="..\NatCruise.Cruise\NatCruise.Cruise.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
