+++ src/NatCruise.Core/Models/TallyPopulationEx.cs
@@ -1,6 +1,7 @@
-﻿namespace NatCruise.Cruise.Models
+﻿namespace NatCruise.Models
 {
-    public class TallyPopulation : TallyPopulation_Base
+    // TODO can this be consolidated into TallyPopulation or broken into a separate model
+    public class TallyPopulationEx : TallyPopulation
     {
         private int _treeCount;
         private int _sumKPI;
@@ -8,13 +9,13 @@
         public int TreeCount
         {
             get { return _treeCount; }
-            set { SetValue(ref _treeCount, value); }
+            set { SetProperty(ref _treeCount, value); }
         }
 
         public int SumKPI
         {
             get { return _sumKPI; }
-            set { SetValue(ref _sumKPI, value); }
+            set { SetProperty(ref _sumKPI, value); }
         }
 
         public bool IsClickerTally { get; set; }
