--- src/NatCruise.Cruise.Test/Data/CuttingUnitDatastore_Test.cs
@@ -1,39 +1,0 @@
-﻿using FluentAssertions;
-using NatCruise.Cruise.Data;
-using NatCruise.Cruise.Models;
-using NatCruise.Test;
-using System.Linq;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace NatCruise.Cruise.Test.Data
-{
-    public class CuttingUnitDatastore_Test : Datastore_TestBase
-    {
-        public CuttingUnitDatastore_Test(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        [Theory]
-        [InlineData("u1", "st3", "st4")]
-        [InlineData("u2", "st4")]
-        public void GetStrataByUnitCode_Test(string unitCode, params string[] expectedStrataCodes)
-        {
-            using (var database = CreateDatabase())
-            {
-                var datastore = new CuttingUnitDatastore(database, CruiseID, TestDeviceInfoService.TEST_DEVICEID);
-
-                var strata = database.Query<Stratum>
-                    ("select * from stratum;").ToArray();
-
-                var stuff = database.QueryGeneric("select * from Stratum;").ToArray();
-
-                var results = datastore.GetStrataByUnitCode(unitCode);
-
-                var strata_codes = results.Select(x => x.StratumCode);
-                strata_codes.Should().Contain(expectedStrataCodes);
-                strata_codes.Should().HaveSameCount(expectedStrataCodes);
-            }
-        }
-    }
-}
