+++ src/NatCruise.Cruise/Data/PlotDataservice.cs
@@ -130,85 +130,6 @@ AND st.Method != '{CruiseMethods.THREEPPNT}';",
 
         #endregion plot
 
-        public IEnumerable<TallyPopulation_Plot> GetPlotTallyPopulationsByUnitCode(string unitCode, int plotNumber)
-        {
-            var cruiseID = CruiseID;
-
-            var tallyPops = Database.Query<TallyPopulation_Plot>(
-@"WITH tallyPopTreeCounts AS (
-    SELECT CruiseID,
-        CuttingUnitCode,
-        StratumCode,
-        SampleGroupCode,
-        SpeciesCode,
-        LiveDead,
-        sum(TreeCount) AS TreeCount,
-        sum(KPI) AS SumKPI
-    FROM TallyLedger AS tl
-    WHERE CuttingUnitCode = @p1 AND CruiseID = @p2
-    GROUP BY
-        CruiseID,
-        CuttingUnitCode,
-        StratumCode,
-        SampleGroupCode,
-        ifnull(SpeciesCode, ''),
-        ifnull(LiveDead, ''))
-
-SELECT
-    tp.Description,
-    tp.StratumCode,
-    st.Method AS StratumMethod,
-    tp.SampleGroupCode,
-    tp.SpeciesCode,
-    tp.LiveDead,
-    tp.HotKey,
-    ifnull(tl.TreeCount, 0) AS TreeCount,
-    ifnull(tl.SumKPI, 0) AS SumKPI,
-    --sum(tl.KPI) SumKPI,
-    sg.SamplingFrequency AS Frequency,
-    sg.MinKPI AS sgMinKPI,
-    sg.MaxKPI AS sgMaxKPI,
-    sg.UseExternalSampler
--- ss.SampleSelectorType == '{CruiseMethods.CLICKER_SAMPLER_TYPE}' AS IsClickerTally
-FROM TallyPopulation AS tp
-JOIN SampleGroup AS sg USING (StratumCode, SampleGroupCode, CruiseID)
--- Left JOIN SamplerState ss USING (StratumCode, SampleGroupCode)
-JOIN Stratum AS st USING (StratumCode, CruiseID)
-JOIN CuttingUnit_Stratum AS cust USING (StratumCode, CruiseID)
-LEFT JOIN tallyPopTreeCounts AS tl
-        ON tl.CuttingUnitCode = cust.CuttingUnitCode
-        AND tl.CruiseID =  tp.CruiseID
-        AND tp.StratumCode = tl.StratumCode
-        AND tp.SampleGroupCode = tl.SampleGroupCode
-        AND ifnull(tp.SpeciesCode, '') = ifnull(tl.SpeciesCode, '')
-        AND ifnull(tp.LiveDead, '') = ifnull(tl.LiveDead, '')
-WHERE cust.CuttingUnitCode = @p1 AND tp.CruiseID = @p2 AND st.Method IN (SELECT Method FROM LK_CruiseMethod WHERE IsPlotMethod = 1)"
-                , new object[] { unitCode, cruiseID }).ToArray();
-
-            foreach (var pop in tallyPops)
-            {
-                pop.InCruise = GetIsTallyPopInCruise(unitCode, plotNumber, pop.StratumCode, cruiseID);
-                pop.IsEmpty = Database.ExecuteScalar<int>("SELECT ifnull(IsEmpty, 0) FROM Plot_Stratum " +
-                    "WHERE CuttingUnitCode = @p1 AND PlotNumber = @p2 AND StratumCode = @p3 AND CruiseID = @p4;",
-                    unitCode, plotNumber, pop.StratumCode, cruiseID) == 1;
-            }
-
-            return tallyPops;
-        }
-
-        private bool GetIsTallyPopInCruise(string unitCode, int plotNumber, string stratumCode, string cruiseID)
-        {
-            return Database.ExecuteScalar<bool?>(
-                "SELECT EXISTS (" +
-                    "SELECT * " +
-                    "FROM Plot_Stratum " +
-                    "WHERE StratumCode = @p1 " +
-                        "AND CuttingUnitCode = @p2 " +
-                        "AND PlotNumber = @p3 " +
-                        "AND CruiseID = @p4);",
-                stratumCode, unitCode, plotNumber, cruiseID) ?? false;
-        }
-
         #region plot stratum
 
         //public IEnumerable<StratumProxy> GetPlotStrataProxies(string unitCode)
