+++ src/NatCruise.Cruise/Services/ITreeBasedTallyService.cs
@@ -1,4 +1,5 @@
 ﻿using NatCruise.Cruise.Models;
+using NatCruise.Models;
 using System.Threading.Tasks;
 
 namespace NatCruise.Cruise.Services
@@ -6,6 +7,6 @@ namespace NatCruise.Cruise.Services
     public interface ITreeBasedTallyService
     {
         Task<TallyEntry> TallyAsync(string unitCode,
-            TallyPopulation pop);
+            TallyPopulationEx pop);
     }
 }
