+++ src/NatCruise.Design/ViewModels/CruiseViewModel.cs
@@ -1,7 +1,4 @@
 ﻿using NatCruise.Data;
-using NatCruise.Data.Abstractions;
-using NatCruise.Design.Data;
-using NatCruise.Design.Models;
 using NatCruise.Design.Validation;
 using NatCruise.Models;
 using System;
