+++ src/NatCruise.Design/ViewModels/StratumTemplateListViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using NatCruise.Data;
-using NatCruise.Data.Abstractions;
 using NatCruise.Design.Data;
 using NatCruise.Design.Models;
 using NatCruise.Models;
